redis_queuing/pyproject.toml
[project]
name = "redis_queuing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
]
